Split backend user feature to sub-directories
