+++ server/src/features/player-assignment/utils/removeInvalidSignupsFromUsers.test.ts
@@ -5,7 +5,8 @@ import { GameModel } from "server/features/game/gameSchema";
 import { mockUser, mockSignup } from "server/test/mock-data/mockUser";
 import { testGame, testGame2 } from "shared/tests/testGame";
 import { removeInvalidGamesFromUsers } from "server/features/player-assignment/utils/removeInvalidGamesFromUsers";
-import { saveSignedGames, saveUser } from "server/features/user/userRepository";
+import { saveUser } from "server/features/user/userRepository";
+import { saveSignedGames } from "server/features/user/signed-game/signedGameRepository";
 
 let mongoServer: MongoMemoryServer;
