+++ server/src/features/player-assignment/utils/removeOverlapSignups.ts
@@ -1,18 +1,16 @@
 import moment from "moment";
 import { logger } from "server/utils/logger";
-import { UserSignup } from "server/typings/result.typings";
+import { UserSignedGames } from "server/typings/result.typings";
 import { User } from "shared/typings/models/user";
-import {
-  findUsers,
-  saveSignedGames,
-} from "server/features/user/userRepository";
+import { findUsers } from "server/features/user/userRepository";
 import { Result } from "shared/typings/models/result";
+import { saveSignedGames } from "server/features/user/signed-game/signedGameRepository";
 
 export const removeOverlapSignups = async (
   results: readonly Result[]
 ): Promise<void> => {
   logger.debug("Find overlapping signups");
-  const signupData: UserSignup[] = [];
+  const signupData: UserSignedGames[] = [];
 
   let users: User[];
   try {
