+++ server/src/features/user/userController.test.ts
@@ -4,8 +4,6 @@ import { ConventionType } from "shared/config/sharedConfig.types";
 import { startTestServer, stopTestServer } from "server/test/utils/testServer";
 import { UserGroup } from "shared/typings/models/user";
 import { getJWT } from "server/utils/jwt";
-import { GroupRequest } from "shared/typings/api/groups";
-import { SaveFavoriteRequest } from "shared/typings/api/favorite";
 
 jest.mock("server/utils/logger");
 
@@ -179,183 +177,6 @@ describe(`POST ${ApiEndpoint.USERS_PASSWORD}`, () => {
         })
         .set("Authorization", `Bearer ${getJWT(UserGroup.USER, "testuser")}`);
       expect(response.status).toEqual(200);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-});
-
-describe(`POST ${ApiEndpoint.FAVORITE}`, () => {
-  test("should return 422 without valid body", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).post(ApiEndpoint.FAVORITE);
-      expect(response.status).toEqual(422);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-
-  test("should return 401 without valid authorization", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    const saveFavoriteRequest: SaveFavoriteRequest = {
-      username: "testuser",
-      favoritedGameIds: [],
-    };
-
-    try {
-      const response = await request(server)
-        .post(ApiEndpoint.FAVORITE)
-        .send(saveFavoriteRequest);
-      expect(response.status).toEqual(401);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-});
-
-describe(`GET ${ApiEndpoint.GROUP}`, () => {
-  test("should return 422 without valid body", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).get(ApiEndpoint.GROUP);
-      expect(response.status).toEqual(422);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-
-  test("should return 401 without valid authorization", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).get(ApiEndpoint.GROUP).query({
-        username: "testuser",
-        groupCode: "1234",
-      });
-      expect(response.status).toEqual(401);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-});
-
-describe(`POST ${ApiEndpoint.GROUP}`, () => {
-  test("should return 422 without valid body", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).post(ApiEndpoint.GROUP);
-      expect(response.status).toEqual(422);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-
-  test("should return 401 without valid authorization", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    const groupRequest: GroupRequest = {
-      groupCode: "1234",
-      isGroupCreator: true,
-      ownSerial: "1234",
-      username: "testuser",
-      leaveGroup: false,
-      closeGroup: false,
-    };
-
-    try {
-      const response = await request(server)
-        .post(ApiEndpoint.GROUP)
-        .send(groupRequest);
-      expect(response.status).toEqual(401);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-});
-
-describe(`POST ${ApiEndpoint.LOGIN}`, () => {
-  test("should return 422 without any parameters", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).post(ApiEndpoint.LOGIN);
-      expect(response.status).toEqual(422);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-
-  test("should return 422 if username is found but password is missing", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).post(ApiEndpoint.LOGIN).send({
-        username: "testuser",
-      });
-      expect(response.status).toEqual(422);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-
-  test("should return 422 if password is found but username is missing", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).post(ApiEndpoint.LOGIN).send({
-        password: "testpass",
-      });
-      expect(response.status).toEqual(422);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-
-  test("should return 200 if password and username are found", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server)
-        .post(ApiEndpoint.LOGIN)
-        .send({ username: "testuser", password: "testpass" });
-      expect(response.status).toEqual(200);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-});
-
-describe(`POST ${ApiEndpoint.SIGNED_GAME}`, () => {
-  test("should return 422 without valid body", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server).post(ApiEndpoint.SIGNED_GAME);
-      expect(response.status).toEqual(422);
-    } finally {
-      await stopTestServer(server, mongoServer);
-    }
-  });
-
-  test("should return 401 without valid authorization", async () => {
-    const { server, mongoServer } = await startTestServer();
-
-    try {
-      const response = await request(server)
-        .post(ApiEndpoint.SIGNED_GAME)
-        .send({
-          signupData: {
-            username: "testuser",
-            selectedGames: [],
-            signupTime: "2019-11-23T08:00:00Z",
-          },
-        });
-      expect(response.status).toEqual(401);
     } finally {
       await stopTestServer(server, mongoServer);
     }
