+++ server/src/test/test-data-generation/generators/createEnteredGames.ts
@@ -1,13 +1,11 @@
 import { faker } from "@faker-js/faker";
 import { logger } from "server/utils/logger";
-import {
-  findUsers,
-  saveEnteredGame,
-} from "server/features/user/userRepository";
+import { findUsers } from "server/features/user/userRepository";
 import { findGames } from "server/features/game/gameRepository";
 import { findSettings } from "server/features/settings/settingsRepository";
 import { shuffleArray } from "server/utils/shuffleArray";
 import { getRandomInt } from "server/features/player-assignment/utils/getRandomInt";
+import { saveEnteredGame } from "server/features/user/entered-game/enteredGameRepository";
 
 export const createEnteredGames = async (): Promise<void> => {
   logger.info(`Generate EnteredGames data`);
